requirements.txt
django==1.11
django-bootstrap3
django-easy-select2
